+++ src/Revision.Domain/Entities/Devices/Device.cs
@@ -1,5 +1,6 @@
 ﻿using Revision.Domain.Commons;
 using Revision.Domain.Entities.Educations;
+using Revision.Domain.Enums;
 
 namespace Revision.Domain.Entities.Devices;
 
@@ -9,7 +10,7 @@ public class Device : Auditable
     public decimal Price { get; set; }
     public bool Glove { get; set; }
     public bool Fragrant { get; set; }
-    public bool IsActive { get; set; }
+    public DeviceStatus Status { get; set; }
 
     public long EducationId { get; set; }
     public Education Education { get; set; }
