--- src/Revision.Service/DTOs/DevicePayments/DevicePaymentUpdateDto.cs
@@ -1,7 +1,0 @@
-﻿namespace Revision.Service.DTOs.DevicePayments;
-
-public class DevicePaymentUpdateDto
-{
-    public decimal Price { get; set; }
-    public long EducationId { get; set; }
-}
