+++ src/Revision.Service/Interfaces/Payments/ITopicPaymentService.cs
@@ -6,8 +6,6 @@ namespace Revision.Service.Interfaces.Payments;
 public interface ITopicPaymentService
 {
     Task<TopicPaymentResultDto> CreateAsync(TopicPaymentCreationDto dto);
-    //Task<TopicPaymentResultDto> UpdateAsync(long id, TopicPaymentUpdateDto dto);
-    //Task<bool> DeleteAsync(long id);
     Task<TopicPaymentResultDto> GetByIdAsync(long id);
     Task<IEnumerable<TopicPaymentResultDto>> GetAllAsync(PaginationParams pagination);
 }
