+++ src/Revision.WebApi/Controllers/Admin/AdminBaseController.cs
@@ -1,10 +1,8 @@
-﻿using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 
-namespace Revision.WebApi.Controllers.Admin
-{
-    [Authorize(Roles = "Admin, SuperAdmin")]
-    public class AdminBaseController : ControllerBase
-    {}
-}
+namespace Revision.WebApi.Controllers.Admin;
+
+[Authorize(Roles = "Admin, SuperAdmin")]
+public class AdminBaseController : ControllerBase
+{ }
