--- src/Revision.WebApi/Controllers/Client/BaseController.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-
-namespace Revision.WebApi.Controllers.Client
-{
-    [Authorize(Roles = "User")]
-    public class BaseController : ControllerBase
-    {}
-}
