+++ src/Revision.WebApi/Revision.WebApi.csproj
@@ -21,8 +21,4 @@
     <ProjectReference Include="..\Revision.Service\Revision.Service.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Controllers\Common\" />
-  </ItemGroup>
-
 </Project>
